% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\docType{methods}
\name{as.filterVCF,VIF-method}
\alias{as.filterVCF,VIF-method}
\title{An S4 methods for object VIF}
\usage{
\S4method{as.filterVCF}{VIF}(object)
}
\description{
Coerce a list to filterVCF object. Some slots can be empty.
}
