% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_class_methods.R
\docType{methods}
\name{as.list.filterVCF,filterVCF-method}
\alias{as.list.filterVCF,filterVCF-method}
\title{An S4 methods for object filterVCF}
\usage{
\S4method{as.list.filterVCF}{filterVCF}(object)
}
\description{
Coerce a filterVCF to list object. If slots are empty they are filled up by object of the right class.
}
