% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\docType{methods}
\name{as.list.VIF,VIF-method}
\alias{as.list.VIF,VIF-method}
\title{An S4 methods for object VIF}
\usage{
\S4method{as.list.VIF}{VIF}(object)
}
\description{
Coerce a VIF object to list object. If slots are empty they are filled up by object of the right class.
}
