% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\docType{methods}
\name{as.VIF,filterVCF-method}
\alias{as.VIF,filterVCF-method}
\title{An S4 methods for object filterVCF}
\usage{
\S4method{as.VIF}{filterVCF}(object)
}
\description{
Coerce a filterVCF to VIF object. Some slots can be empty.
}
