% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_functions.R
\name{bootstrap_VIF}
\alias{bootstrap_VIF}
\title{Create index of bootstraped vif object}
\usage{
bootstrap_VIF(vif, nb_bootstrap = 200, bootstrap_by = "region",
  seed = NULL)
}
\arguments{
\item{vif}{a vif object}

\item{nb_bootstrap}{Number of bootstrap to create index}

\item{bootstrap_by}{Either a Granges, "region" or a numeric indicative of windows size in "bp"}

\item{seed}{The seed to perform the resampling, if NULL the seed is taken randomly and store in bootSeed(vif)}
}
\value{
A vif object with the slot @bootstrap filled with :
	 \item{bootstrap matrix}{a bootstrap matrix with the index of each block. The matrix has a dimension of nb_block X nb_bootstrap. It can be retrieved using bootstrapMat(vif)}
  \item{number of bootstrap}{retrieved using Nboot(vif)}
  \item{type of bootstrap}{retrieved with bootBy(vif)}
  \item{indexes}{the indexes for the vcf, invariable and fixed. If divergenceare present they al get boostrap with the same seeds making it comparable among bootstrap.}
}
\description{
This function does not produce a bootstrap of the original vif object as it
would be heavy on memmory to copy the entire vcf,invariable, fixed and divergence part of
the vif object. Instead it creates index of position and bootstrap those
indexes. The latter can be then used with get_bootstrap_VIF to produce the
bootstrapped dataset.
}
\details{
The bootstrap is a resample with replacement and be conducted using
different block.

The user can provide a GRange object with blocks that will be intersect
with the vif object, boostrapping with the same see the snp, invariable and
fixed sites ' If a region was already define the bootstrap_by can be "region"
and the function look in the vif object using bootGRange(vif) to get the
GRange object. This is useful if you already bootstrapped and then remove
some samples. ' Finally if bootstrap_by is a numeric, it define the windows
size in bp to create blocks. The functions looks up the reference in
reference(vif) and then create non-overlapping windows on which it perfomrs
the bootstrap
}
\examples{

vif_bootstraped <- bootstrap_VIF(vif, nb_bootstrap = 200, bootstrap_by = 20000)
head(bootstrapMat(vif_bootstraped))

}
\author{
~~Benjamin Laenen~~
}
\keyword{vif}
