% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_filters.R
\name{change_gt_by_DP_filter}
\alias{change_gt_by_DP_filter}
\title{Modify the genotype in the vcf based on depth}
\usage{
change_gt_by_DP_filter(vcf, min_DP = 10, max_DP = 200, ...)
}
\arguments{
\item{vcf}{A vcfR object}

\item{min_DP}{The minimum depth threshold to set genotype as missing}

\item{max_DP}{The minimum depth threshold to set genotype as missing}

\item{...}{list of options in the form a list object: opt. Only opt$verbose
is used here. Default is true.}
}
\value{
A vcfR object with genotype as missing if not passing filters.
}
\description{
This function extract the depth of coverage for each sites and each
individuals and set the genotype as missing (./.) if DP is below the min or
above the max threshold for depth. The fied DP must be present in the FORMAT
column. Note that the number of missing per sites will be modified by this filters.
}
\note{
~~further notes~~
}
\examples{

vcf_modified_dp <- change_gt_by_DP_filter(vcf, min_DP=10, max_DP=200)

}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},

Other filters: \code{\link{change_gt_by_GQ_filter}},
  \code{\link{create_filter_META}},
  \code{\link{create_filter_all_het}},
  \code{\link{create_filter_fix_het_contiguous_in_pop}},
  \code{\link{create_filter_fix_het}},
  \code{\link{create_filter_fixed_sites}},
  \code{\link{create_filter_high_DP_standardized}},
  \code{\link{create_filter_missing}},
  \code{\link{create_filter_repeats_in_windows}},
  \code{\link{create_indel_filter}},
  \code{\link{create_invariant_filter}},
  \code{\link{missing_per_individual}}
}
\author{
~~Benjamin Laenen~~
}
\concept{filters}
\keyword{~GATK}
\keyword{~filters}
