% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_filters.R
\name{change_gt_by_GQ_filter}
\alias{change_gt_by_GQ_filter}
\title{Modify the genotype in the vcf based on genotype quality}
\usage{
change_gt_by_GQ_filter(vcf, invariable = FALSE, GQ_threshold = 5, ...)
}
\arguments{
\item{vcf}{A vcfR object}

\item{invariable}{If the vcf is composed of invariable sites, GATK use RGQ instead of GQ and invariable must be set to TRUE.}

\item{GQ_threshold}{The minimum genotype quality threshold to set genotype as missing. If "auto" set the threshold based on the 5 percent quartile over all sites}

\item{...}{list of options in the form a list object: opt. Only opt$verbose
is used here. Default is true.}
}
\value{
A vcfR object with genotype as missing if not passing filters.
}
\description{
This function extract the genotype quality scores for each sites and each '
individuals and set the genotype as missing (./.) if GQ or RGQ is below the
min threshold for genotype quality. The fied GQ for vairant and RGQ for
invariant must be present in the FORMAT column. If "auto" is used for
threshold the 5 percent quartile is used as a lower threshold for GQ/RGQ.
Note that the number of missing per sites will be modified by this filters.
}
\examples{

vcf_modified_GQ <- change_gt_by_GQ_filter(vcf, GQ_threshold=auto)
vcf_modified_GQ5 <- change_gt_by_GQ_filter(vcf, GQ_threshold=5)

vcf_modified_RGQ <- change_gt_by_GQ_filter(vcf_invariant, invariable = TRUE, GQ_threshold=5)

}
\references{

}
\seealso{
Other filters: \code{\link{change_gt_by_DP_filter}},
  \code{\link{create_filter_META}},
  \code{\link{create_filter_all_het}},
  \code{\link{create_filter_fix_het_contiguous_in_pop}},
  \code{\link{create_filter_fix_het}},
  \code{\link{create_filter_fixed_sites}},
  \code{\link{create_filter_high_DP_standardized}},
  \code{\link{create_filter_missing}},
  \code{\link{create_filter_repeats_in_windows}},
  \code{\link{create_indel_filter}},
  \code{\link{create_invariant_filter}},
  \code{\link{missing_per_individual}}
}
\author{
~~Benjamin Laenen~~
}
\concept{filters}
\keyword{~GATK}
\keyword{~filters}
