% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_filters.R
\name{create_filter_high_DP_standardized}
\alias{create_filter_high_DP_standardized}
\title{Create a filter using small scale normalized DP}
\usage{
create_filter_high_DP_standardized(reference, vcf, threshold = 2,
  threshold_CI = 4, windows_size = 1000, overlapping = 1000,
  percent_passing_filter = 0.95, ...)
}
\arguments{
\item{reference}{The path to the reference genome or a DNAStringSet object with the reference.}

\item{vcf}{A vcfR object}

\item{threshold}{= 2 value(s) for the median normalized DP threshold}

\item{threshold_CI}{= 4 value(s) for the 95 quartile normalized DP threshold}

\item{windows_size}{Size of fixed windows}

\item{percent_passing_filter}{= 0.95 determine the percentage of samples required to pass the threshold of normalized coverage to call the windows as rejected}

\item{...}{list of options in the form a list object: opt. Only opt$verbose
is used here. Default is true.}

\item{overlapping=1000}{sliding window parameter}
}
\value{
a list of logical filter of length @vcf
}
\description{
These filters are designed to try to deal with repetitive genome. Region
with repeats will often have higher depth due to read mapping to two or more
regions. We use short windows (recommended between 1000 and 5000) to
calculate the depth for each samples for each snps in the VCFfile. The depth
is nomralized for each sample by the median depth over all snps. This enable
to compare samples with different coverage and apply relative threshold
instead of raw value of depth. Then two kind of rules are applied. First a
windows is removed if the mean normalized depth is higher than the threshold
and the mean upper 95 quartile is above the CI threshold. For example if the
a sample has median depth of 30X, a windows with 60X and a 95 quartile of 90X
will have a normalized median depth of 2 and 95 quartile of 3. Taking the
mean over samples for this windows might give an value of normalized median
depth of 2.2 and 95 quartile of 3.2. We than apply the threshold (typically 2
and 4) 2.2 > 2 or 3.2 > 4, hence the snps in this windows are filtered. The
second rule is more stringent and remove a windows if any sample is over the
thresholds for a windows. The pipeline use the first rule with the first
value specified in the option --filter_high_DP_standardized#'
}
\examples{

##opt <- initialise_option()
##reference <- "Crubella.fa"
##repeat_bed_file <- "repeats.bed"
##filter <- create_filter_high_DP_standardized(reference, vcf, threshold = 2, threshold_CI = 4, windows_size=1000, overlapping=1000, opt)
##sum(width(filter))

##reference <- readDNAStringSet("Crubella.fa")
##repeat_bed_file <- bed2Grange("repeats.bed")
##filter <- create_filter_high_DP_standardized(reference, vcf, threshold = 2, threshold_CI = 4, windows_size=1000, overlapping=1000, opt)
##sum(width(filter))

}
\references{

}
\seealso{
Other filters: \code{\link{change_gt_by_DP_filter}},
  \code{\link{change_gt_by_GQ_filter}},
  \code{\link{create_filter_META}},
  \code{\link{create_filter_all_het}},
  \code{\link{create_filter_fix_het_contiguous_in_pop}},
  \code{\link{create_filter_fix_het}},
  \code{\link{create_filter_fixed_sites}},
  \code{\link{create_filter_missing}},
  \code{\link{create_filter_repeats_in_windows}},
  \code{\link{create_indel_filter}},
  \code{\link{create_invariant_filter}},
  \code{\link{missing_per_individual}}
}
\author{
~~Benjamin Laenen~~
}
\concept{filters}
\keyword{filter}
