% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_filters.R
\name{create_filter_repeats_in_windows}
\alias{create_filter_repeats_in_windows}
\title{Create filter for percentage of repeats in windows}
\usage{
create_filter_repeats_in_windows(reference, repeat_bed_file, vcf,
  windows_size = 20000, threshold = 0.5, ...)
}
\arguments{
\item{reference}{The path to the reference genome or a DNAStringSet object with the reference.}

\item{repeat_bed_file}{The path to the bedfile or a GRanges object}

\item{vcf}{A vcfR object}

\item{windows_size}{Size of fixed windows}

\item{threshold}{The percentage above which the window is reported in the filter.}

\item{...}{list of options in the form a list object: opt. Only opt$verbose
is used here. Default is true.}
}
\value{
A GRanges object with the windows to remove (not passing the threshold)
}
\description{
The function is intented to remove windows in the genome with more than a
percentage of sites consiting of repeats. In practice the bed file defining
the intervals to consider can be any bedfile. For example a bedfile defining
TE can be used to detect windows with a certain percentage of TE.
}
\note{
~~further notes~~
}
\examples{

##opt <- list(verbose=TRUE)
##reference <- "Crubella.fa"
##repeat_bed_file <- "repeats.bed"
##filter <- create_filter_repeats_in_windows(reference, repeat_bed_file, vcf, windows_size=20000, threshold=0.5, opt)
##sum(width(filter))

##reference <- readDNAStringSet("Crubella.fa")
##repeat_bed_file <- bed2Grange("repeats.bed")
##filter <- create_filter_repeats_in_windows(reference, repeat_bed_file, vcf, windows_size=20000, threshold=0.5, opt)
##sum(width(filter))

}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},

Other filters: \code{\link{change_gt_by_DP_filter}},
  \code{\link{change_gt_by_GQ_filter}},
  \code{\link{create_filter_META}},
  \code{\link{create_filter_all_het}},
  \code{\link{create_filter_fix_het_contiguous_in_pop}},
  \code{\link{create_filter_fix_het}},
  \code{\link{create_filter_fixed_sites}},
  \code{\link{create_filter_high_DP_standardized}},
  \code{\link{create_filter_missing}},
  \code{\link{create_indel_filter}},
  \code{\link{create_invariant_filter}},
  \code{\link{missing_per_individual}}
}
\author{
~~Benjamin Laenen~~
}
\concept{filters}
\keyword{filter}
