% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_filters.R
\name{create_invariant_filter}
\alias{create_invariant_filter}
\title{Create filter for invariant}
\usage{
create_invariant_filter(vcf, gt = NULL, ...)
}
\arguments{
\item{vcf}{A vcfR object}

\item{gt}{genotype matrix created with extract.gt or extract_gt_ff}

\item{...}{list of options in the form a list object: opt. Only opt$verbose
is used here. Default is true.}
}
\value{
logical vector of sizes length(vcf) with indel sites as TRUE
}
\description{
This function check the vcf for invariant sites. Invariant sites are either
sites that have a "." in the ALT field or if a genotype matrix is given sites
that are all homozygous for the reference allele or all homozygous for the
reference allele and missing. In practice it can be used to separate SNPs
from invariant.
}
\note{
~~further notes~~
}
\examples{

##opt <- list(verbose=TRUE)
##filter <- create_invariant_filter(vcf, opt)
##vcf_snp <- vcf[!filter]
##vcf_inv <- vcf[filter]
##gt <- extract_gt_ff(vcf_snp)
##filter_invariant_from_genotype <- create_invariant_filter(vcf_snp,gt = gt, opt)
##sum(filter)
##sum(filter_invariant_from_genotype)

}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},

Other filters: \code{\link{change_gt_by_DP_filter}},
  \code{\link{change_gt_by_GQ_filter}},
  \code{\link{create_filter_META}},
  \code{\link{create_filter_all_het}},
  \code{\link{create_filter_fix_het_contiguous_in_pop}},
  \code{\link{create_filter_fix_het}},
  \code{\link{create_filter_fixed_sites}},
  \code{\link{create_filter_high_DP_standardized}},
  \code{\link{create_filter_missing}},
  \code{\link{create_filter_repeats_in_windows}},
  \code{\link{create_indel_filter}},
  \code{\link{missing_per_individual}}
}
\author{
~~Benjamin Laenen~~
}
\concept{filters}
\keyword{filter}
