% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_class_methods.R
\docType{class}
\name{filterVCF-class}
\alias{filterVCF-class}
\title{An S4 class to filterVCF object.}
\description{
An S4 class to filterVCF object.
}
\section{Slots}{

\describe{
\item{\code{vcf}}{The original vcf with only snp of class "vcfR",}

\item{\code{vcf_filtered}}{The filtered vcf of class "vcfR",}

\item{\code{vcf_inv}}{The list of invariant sites of class "GRanges",}

\item{\code{vcf_inv_filtered}}{The list of invariant filtered sites of class "GRanges",}

\item{\code{fixed_inv_Grange}}{The list of invariant fixed sites of class"GRanges",}

\item{\code{reference}}{A DNAStringSet object with the reference used to create the object,}

\item{\code{master_filter}}{All sites too filter for snps. A logical vector with the same length as the @vcf,}

\item{\code{master_filter_inv}}{All sites too filter for invariant. A logical vector with the same length as the @vcf_inv,}

\item{\code{filters}}{A list of all the filters applied to snp. List of logical vector with length equal to @vcf,}

\item{\code{filters_inv}}{A list of all the filters applied to invariants. List of logical vector with length equal to @vcf,}

\item{\code{removed_sites}}{All sites removed in snp of class "GRanges",}

\item{\code{removed_sites_inv}}{All sites removed in invariant of class "GRanges",}

\item{\code{bed_GRange_merged}}{Merged sites from the bedfiles inputed for filtering of class "GRanges,}

\item{\code{windows_with_half_repeats_to_remove}}{Windows with 50 percent of repeats of class "GRanges",}

\item{\code{fix_het_pop}}{Sites that are fixed heterozygotes in the selected population of class "GRanges",}

\item{\code{filter_DP_repeats}}{A list of filter based on local standardised depth. A list of logical vector of length @vcf,}

\item{\code{stats_vcf_filtered}}{Unused for the moment. Class "character",}

\item{\code{opt}}{List of option used by the pipeline filterVCF.R,}
}}

