% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_class_methods.R
\name{filterVCF}
\alias{filterVCF}
\title{Create an empty filterVCF S4 object}
\usage{
filterVCF()
}
\value{
empty filterVCF object
}
\description{
This function is used to initialize an filterVCF S4 object with all the
necessary slots accesible with @. The slots can be listed by using
slotNames(filterVCF()). Initializing with this function is equivalent to
new("filterVCF") an ensure that the slots have the correct class.
}
\note{
~~further notes~~
}
\examples{

data <- filterVCF()
data
slotNames(data)

}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},
}
\author{
~~Benjamin Laenen~~
}
\keyword{filterVCF}
