% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_VCF_utilities.R
\name{get_alternate_GRange}
\alias{get_alternate_GRange}
\title{Get a genomic range with REF and ALT bases}
\usage{
get_alternate_GRange(filterVCF)
}
\arguments{
\item{filterVCF}{a filterVCF object}
}
\value{
A genomicRange object with position from the vcf and REF and ALT
base as metadata
}
\description{
The functions extract the REF and ALt column present in the original vcf
file and report the base. It ignores the invariant (ALT = "*") but it
includes the fixed sites.
}
\examples{

ref_alt <- get_alternate_GRange(filterVCF)
summary(elementMetadata(ref_alt))
}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{~utilities}
