% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_functions.R
\name{get_bootstrap_freq}
\alias{get_bootstrap_freq}
\title{Get one bootsrap replicate of frequency object}
\usage{
get_bootstrap_freq(freq, vif, bootstrap_matrix = bootstrapMat(vif),
  index = 1)
}
\arguments{
\item{freq}{= NULL provide a frequency object constructed using get_frequency(). This is meant to speed the retrieving of the SFS as the main speed bottleneck is getting the allele frequencies.}

\item{vif}{a VIF object}

\item{index}{index of a single bootstrap to extract, should be lower than NBoot(vif)}
}
\value{
a boostraped frequency object
}
\description{
This function is meant for internal speed up in make_input_DFE and should only be used by experienced user
}
\examples{

#create the index for 20kb windows bootstrap
vif <- bootstrap_VIF(vif, nb_bootstrap = 200, bootstrap_by = 20000)
freq <- get_frequency(vif)
#get the first replicate
boot_freq <- list()
boot_freq[[1]] <- get_bootstrap_freq(freq, vif, index = 1 )

}
\references{
put references to the literature/web site here
}
\seealso{
objects to See Also as \code{\link{help}},
}
\author{
~~Benjamin Laenen~~
}
\keyword{~kwd1}
\keyword{~kwd2}
