% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_convert.R
\name{Grange2bed}
\alias{Grange2bed}
\title{Save GRanges to BED file}
\usage{
Grange2bed(Grange_filter, name = "", outputdir = ".", output_name,
  merge = TRUE, keep_extra_col = FALSE, ...)
}
\arguments{
\item{Grange_filter}{the logical vector created by a filter function}

\item{name}{the name of the filter}

\item{outputdir}{= "." the output directory to save the file to}

\item{output_name}{the name of the BED file to save}

\item{...}{<what param does>}

\item{vcf_bed_GRange}{A GRange object with the position of the VCF corresponding to the filter}
}
\value{
Write to file
}
\description{
This function write a BED file from a GRanges object. It is a wrapper that
uses internally rtracklayer export to export to BED. Note that the intervals
are merged before saving to reduce file size. If you one to keep all interval
you should use the export from rtracklayer directly.
}
\note{
~~further notes~~
}
\examples{

# Exporting the position of the vcf in a BED format
vcf_bed_GRange <- vcf2Grange(vcfR)
Grange2bed(vcf_bed_GRange, "vcf", vcf_bed_GRange, output_name = "vcf.bed")


}
\references{

}
\seealso{
Other utilities: \code{\link{Grange2filter}},
  \code{\link{bed2Grange}}
}
\author{
~~Benjamin Laenen~~
}
\concept{utilities}
\keyword{~utilities}
