% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_convert.R
\name{Grange2filter}
\alias{Grange2filter}
\title{Transform GRange to filter}
\usage{
Grange2filter(Grange, vcf_bed_GRange, ...)
}
\arguments{
\item{Grange}{a Genomic range object}

\item{vcf_bed_GRange}{A GRange object with the position of the VCF corresponding to the filter}

\item{...}{<what param does>}
}
\value{
a logical vector of the size of nrow(vcf)
}
\description{
This function takes a GRanges object and the GRanges from the vcf to create
a logical vector matching the position of the vcf.
}
\note{
~~further notes~~
}
\examples{

vcf_bed_GRange <- vcf2Grange(vcfR)
filter <- Grange2filter(Grange, vcf_bed_GRange)
vcf[!filter]

}
\references{

}
\seealso{
Other utilities: \code{\link{Grange2bed}},
  \code{\link{bed2Grange}}
}
\author{
~~Benjamin Laenen~~
}
\concept{utilities}
\keyword{~utilities}
