% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_VCF_utilities.R
\name{initialise_option}
\alias{initialise_option}
\title{Initialize the option to run filterVCF.R}
\usage{
initialise_option()
}
\value{
A list of options
}
\description{
Should not be used by end user.
}
\details{
Set all the default of opt when initializing 
an empty filterVCF object
}
\examples{

opt <- parse_args(OptionParser(option_list=option_list))

}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{~utilities}
