% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\docType{methods}
\name{initialize,VIF-method}
\alias{initialize,VIF-method}
\title{An S4 methods for object list}
\usage{
\S4method{initialize}{VIF}(.Object, ...)
}
\description{
Initialise an empty VIF object
}
