% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_functions.R
\name{main_filter_VCF}
\alias{main_filter_VCF}
\title{Main wrapper to filter a VCF file}
\usage{
main_filter_VCF(vcf_file, ...)
}
\arguments{
\item{vcf_file}{path to vcf file}

\item{...}{options from initialise_option(). This parameter regulates all the filtering steps that will be applied to the data.}
}
\value{
a filterVCF object.
}
\description{
This function is the main core of the pipeline filterVCF.R. It will perform
a series of filtering steps depending of the options provided in a opt list
that can be initiated with initialise_option(). This function is not expected
to be used directly by the user which should first run the pipeline
filterVCF.R on a complete non filtered VCF (including invariant).
}
\details{
see the help of filterVCF.R for more detail
Rscript filterVCF.R --help
}
\examples{

opt <- parse_args(OptionParser(option_list=option_list))
main_filter_VCF("my.vcf",  opt)

}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{~main}
