% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_repolarization.R
\name{make_GRange_freq_base_Keightley}
\alias{make_GRange_freq_base_Keightley}
\title{Create input to run P.Keightley 'est-sfs'}
\usage{
make_GRange_freq_base_Keightley(vif, approximate_freq = TRUE, ...)
}
\arguments{
\item{vif}{a VIF object}

\item{approximate_freq}{Should the frequency by approximated as explained above?}
}
\value{
A GRange object with metadata suitable to create an input for est-sfs. The values associated with each sitesare the following
\item{freq }{allele frequency of the alternates alleles} %% 
\item{nsample }{total number of chromosomes sampled} %% ...
\item{REF }{base of the REF} %%
\item{ALT }{base of the ALT} %%
\item{OUT1 }{base of the first outgroup} %%
\item{OUT2 }{base of the second outgroup} %%
}
\description{
This function takes a vif object and create an input to run P.Keightly
program that aims at estimating the unfolded SFS using two outgroups in a
maximum likelihood framework.
It is expected that the slot @divergence of the vif object is 
filled and two outgroups are present.
The program est-sfs require that all sites have the same number of
alleles. If we don't want to loose too much, we can approximate the
frequency with freq / (nb sampled allele * max nb of sampled alleles). This
approxiation is robust if there is not too much missing data.
}
\details{
Below is the descriptoin from the manual on the input specifications:
# The format is the following:
#     There are 12 space-separated columns, 4 for the focal species and 4
# for each of the two outgroups. If the single outgroup program is used,
# the last 4 columns are ignored, but must be present (they can just be
# a duplication of outgroup 1). Each line in the input file gives the
# count of alleles at a site. The order of bases is A C G T.

# For example, the first line in the example data file is:

# 20 0 0 0                0 0 0 1 0 0 0 1

# At this site, all of the n = 20 copies sampled are A in the focal
# species. In the outgroups, a single copy has been sampled, and both
# are T. All sites must have the same number of copies sampled in the
# focal species and one copy sampled in each of the outgroups.
}
\examples{

input_est-sfs <- make_GRange_freq_base_Keightley(vif_object)
input_est-sfs

}
\author{
~~Benjamin Laenen~~
}
\keyword{~est-sfs}
