% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_functions.R
\name{make_input_DFE}
\alias{make_input_DFE}
\title{Wrapper to create input for DFE-alpha or polyDFE}
\usage{
make_input_DFE(vif, nonsyn, syn, name, output = c("DFE-alpha",
  "polyDFE"), folded = FALSE, downsample_size = NULL,
  nb_bootstrap = NULL, bootstrap_by = "region", mc.cores = 4,
  use_repol = TRUE, remove_polym_from_div = TRUE, stat = TRUE,
  bootstrapped_vif_syn = NULL, bootstrapped_vif_nonsyn = NULL,
  bootstrapped_freq_syn = NULL, bootstrapped_freq_nonsyn = NULL)
}
\arguments{
\item{vif}{a VIF object}

\item{nonsyn}{A Granges with selected sites, A path to a bed file with selected sites or a VIF object containing only selected sites}

\item{syn}{A Granges with neutral sites, A path to a bed file with neutral sites or a VIF object containing only neutral sites}

\item{name}{Root name to give in the input}

\item{output}{= c("DFE-alpha", "polyDFE") type of output}

\item{folded}{= FALSE fold the SFS?}

\item{downsample_size}{= NULL integer indicating the size of the downsampling}

\item{nb_bootstrap}{= NULL Number of desired bootstrap. If NULL, no bootstrap is performed (can be usefull for quicj debugging)}

\item{mc.cores}{= 4 number of core to use.}

\item{use_repol}{= TRUE Should a repolarizion using the outgroup be performed?}

\item{remove_polym_from_div}{= TRUE Remove the contribution of polymorphism from the divergence}

\item{stat}{currently unused}

\item{bootstrapped_vif_syn}{= NULL advanced option, provided a list with boostrapped vif for neutral sites to improve perfomance when calling the function multiple times.}

\item{bootstrapped_vif_nonsyn}{= NULL advanced option, see above}

\item{bootstrapped_freq_syn}{= NULL advanced option, see above}

\item{bootstrapped_freq_nonsyn}{= NULL advanced option, see above}

\item{bootstrap_by="region"}{Should the bootstrap be perform by region given in bootGRange(vif) or by windows. If windows is desired an integer representing windows size in bp is required.}
}
\value{
a large list with intersect element and output
\item{SFS }{unfolded SFS} %% 
\item{fSFS }{folded SFS} %% 
\item{down_SFS }{downsampled unfolded SFS} %% 
\item{down_fSFS }{downsampled folded SFS} %% 
\item{D }{number of divergent sites between ingroup and outgroup(s)} %% 
\item{lD }{total number of divergence} %% 
\item{vif_syn }{vif object for neutral sites} %% 
\item{vif_nonsyn }{vif object for selected sites} %% 
\item{VIF_out }{The actual input for DFE-alpha or polyDFE as a character vector} %% 
\item{sfs_syn }{SFS list for neutral sites} %% 
\item{sfs_nonsyn }{SFS list for selected sites} %% 
\item{bootstrapped_sfs_syn }{list of bootstrapped neutral SFS} %% 
\item{bootstrapped_sfs_nonsyn }{list of bootstrapped selected SFS} %% 
\item{bootstrapped_vif_syn }{ list of bootstrapped neutral VIF object} %% 
\item{bootstrapped_vif_nonsyn }{ list of bootstrapped selected VIF object} %% 
\item{bootstrapped_freq_syn }{ list of bootstrapped neutral frequency object} %% 
\item{bootstrapped_freq_nonsyn }{ list of bootstrapped selected frequency object} %%
}
\description{
This function is intented to automatically create input for the program
DFE-alpha or polyDFE using information in the VIF object and GRange object
for selected and neutral sites. If divergence is present in the VIF object,
repolarization can be performed using either parsimony or ML. Note that for
ML the function use the software of P. Keightley 'est-sfs' which should be
present in your $PATH and be callable from a system() call.
}
\examples{

}
\author{
~~Benjamin Laenen~~
}
\keyword{dfe}
