% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_optionParser.R
\name{parse_filter_depth}
\alias{parse_filter_depth}
\title{parse options for filter individual depth}
\usage{
parse_filter_depth(opt_filter_depth)
}
\arguments{
\item{opt_filter_depth}{Character string, see help of filterVCF.R pipeline for more info}
}
\value{
list of min and max depth threshold
}
\description{
Intented to be use with the pileline filterVCF.R
}
\examples{

Parse the option from a call to 
option_list <- initialise_option()
opt <- parse_args(OptionParser(option_list=option_list))
opt$filter_depth <- "10,200"
parse_filter_depth(opt$filter_depth)

}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{parser}
