% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_optionParser.R
\name{parse_filter_high_DP_standardized}
\alias{parse_filter_high_DP_standardized}
\title{parse options for filter using normalized depth
Intented to be use with the pileline filterVCF.R}
\usage{
parse_filter_high_DP_standardized(opt_filter_high_DP_standardized)
}
\arguments{
\item{opt_filter_high_DP_standardized}{Character string, see help of filterVCF.R pipeline for more info}
}
\value{
list of of filter threshold and windows size to be used during the filtering
}
\description{
parse options for filter using normalized depth
Intented to be use with the pileline filterVCF.R
}
\examples{

Parse the option from a call to 
option_list <- initialise_option()
opt <- parse_args(OptionParser(option_list=option_list))
opt$filter_high_DP_standardized <- "(2 1.2 1.5 2.2),4,1000,1000"
parse_filter_high_DP_standardized(opt$filter_high_DP_standardized)


}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{parser}
