% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_optionParser.R
\name{parse_opt_popfile}
\alias{parse_opt_popfile}
\title{parse options for file with population assignment}
\usage{
parse_opt_popfile(opt_popfile)
}
\arguments{
\item{opt_popfile}{path to a tab-separated popfile, see help of filterVCF.R pipeline for more info}
}
\value{
dataframe with sample names and population
}
\description{
Intented to be use with the pileline filterVCF.R
}
\examples{

Parse the option from a call to 
option_list <- initialise_option()
opt <- parse_args(OptionParser(option_list=option_list))
opt$popfile <- "popfile.tsv"
parse_opt_popfile(opt$popfile)

}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{parser}
