% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_GRange_utilities.R
\name{plotRanges}
\alias{plotRanges}
\title{Plot a Grange object}
\usage{
plotRanges(x, xlim = x, main = deparse(substitute(x)), col = "black",
  sep = 0.5, colorRange = FALSE, colorRamp = c("cadetblue1",
  "brown4"), exclude.column = NULL, ...)
}
\arguments{
\item{x}{A GRange}

\item{colorRange}{= FALSE Should the value associated to each GRange be represented by a gradient}

\item{colorRamp}{= c("cadetblue1","brown4") Start and end of the color gradient}

\item{exclude.column}{= NULL Index of the column to exclude from plot}

\item{...}{paremeter pass to rect() except col and border}

\item{xlim=x}{region to plot, default spans the min and max of the GRange}

\item{main=deparse(substitute(x))}{Plot title}

\item{col="black"}{single value or list of color the length of the GRanges}

\item{sep=0.5}{Not in use}
}
\description{
Simple representation of GRange interval using rectangle. If values are
associated with each range multiple track can be visualized at the same time.
}
\examples{


}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},
}
\author{
~~Benjamin Laenen~~
}
\concept{plot}
\keyword{~plot}
