% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_class_methods.R
\docType{methods}
\name{repair,filterVCF-method}
\alias{repair,filterVCF-method}
\title{An S4 methods for object filterVCF}
\usage{
\S4method{repair}{filterVCF}(object)
}
\description{
Repair an object by filling the missing slot with empty object. Function for compatibility if the definition changes between version.
}
