% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\docType{methods}
\name{repairVIF,VIF-method}
\alias{repairVIF,VIF-method}
\title{An S4 methods for object VIF}
\usage{
\S4method{repairVIF}{VIF}(object)
}
\description{
Repair an object by filling the missing slot with empty object. Function for compatibility if the definition changes between version.
}
