% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_repolarization.R
\name{repolarize_vcf}
\alias{repolarize_vcf}
\title{Polarize a vcfR}
\usage{
repolarize_vcf(vcf, repolKey, methods = "ML")
}
\arguments{
\item{vcf}{vcfR object}

\item{repolKey}{a repolarization key create using make_repol_key()}

\item{methods}{"ML" OR "parsimony". Which methods to use to repolarize? Note that the repolKey must have been create with the methods in ordor to use it (logical!)}
}
\value{
a polarized vcfR object
}
\description{
Internal function used in repolarize_VIF(). Only experienced users. NOte that invariant and fixed will not be polarized here.
}
\examples{

repolKey(vif_object) <- make_repol_key(vif_object, methods = c("ML", "parsimony"))
polarized_VCF  <- repolarize_VIF(vcf(vif_object), methods = "ML")


}
\references{
put references to the literature/web site here
}
\author{
~~Benjamin Laenen~~
}
\keyword{~repolarization}
