% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_repolarization.R
\name{run_Keightley_estsfs}
\alias{run_Keightley_estsfs}
\title{Wrapper to run est-sfs from P.Keightley}
\usage{
run_Keightley_estsfs(ACGT, Model = c("JC", "kimura", "rate6"),
  nrandom = 5, noutgroups = 2, bin_path = "est-sfs",
  write_files = FALSE, name = "", ...)
}
\arguments{
\item{ACGT}{output from make_GRange_freq_base_Keightley()}

\item{Model}{= "JC", "kimura", "rate6". Sequence evolution model to use. A LRT can be conducted ad hoc to shoose the best model}

\item{nrandom}{= 5 number of random start}

\item{noutgroups}{= 2 Number of outgroup (fixed at 2 for the moment)}

\item{bin_path}{= "est-sfs" location of the executable}

\item{name}{= "" common root name for output}

\item{...}{options (NOT USED)}

\item{write_files=FALSE}{Should a full output be written to a folder}
}
\value{
A data.frame with the same number of sites as the input Grange. It contains the probability that the major allele is ancestral (see est-sfs manual for more detail)
}
\description{
This function runs est-sfs from P.Keightley. The program need to be
installed and and be callable using a system() call. This has only been test
on UNIX (mac and linux). This is used internally in the larger function
make_repol_key() that should be used to repolarize an vif or filterVCF
object. However, this wrapper function allows more freedom in the model.
}
\details{
The aim of the function is to return the file with the ancestral
probability for each sites but full output from the program can also be
available if write_files = TRUE
}
\examples{

input_est-sfs <- make_GRange_freq_base_Keightley(vif_object)
run_Keightley_estsfs(input_est-sfs, Model = "kimura", bin_path = "est-sfs", write_files = TRUE)

}
\author{
~~Benjamin Laenen~~
}
\keyword{~est-sfs}
