% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\docType{methods}
\name{show,VIF-method}
\alias{show,VIF-method}
\title{An S4 methods for object filterVCF}
\usage{
\S4method{show}{VIF}(object)
}
\description{
Print information on the current object.
}
