% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vcf_utilities.R
\name{split_vcf_by_chrom}
\alias{split_vcf_by_chrom}
\title{Utility to split a VCF file by chromosome}
\usage{
split_vcf_by_chrom(vcf_file, outputdir, ...)
}
\arguments{
\item{vcf_file}{path to a vcf file (gziped or not)}

\item{outputdir}{Output directory to save the splitted vcf}
}
\value{
a character vector with the names of the splitted files.
}
\description{
This function read the VCF file using a quick read.line function and save
splitted vcf file by chromosome on disk. It calls system to gunzip/unpigz and
is only tested on UNIX system. It is used internally in the pipeline
filterVCF.R. If you encounter problem, try to gunzip the vcffile first and then split.
}
\examples{

opt = initialise_option()
opt$gzip_program <- "gzip"
split_vcf_by_chrom("my.vcf", "split_out", opt)


}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{~utilities}
