% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vcf_utilities.R
\name{split_vcfR}
\alias{split_vcfR}
\title{Split a vcfR by GRange}
\usage{
split_vcfR(vcfR, Grange)
}
\arguments{
\item{vcfR}{A vcfR object}

\item{Grange}{A GRange object}
}
\value{
a list of vcfR object the size of length(Grange)
}
\description{
Use a Grange object as a factor to split a vcfR object. Each ranges will be intersect with the vcfR.
}
\examples{

#Split a vcfR by gene
#' gff <- import("my_species.gff3", format = "gff")
gff <- gff[values(gff)$type == "gene"]
vcf_by_gene <- split_vcfR(vcfR, gff)

}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},
}
\author{
~~Benjamin Laenen~~
}
\keyword{~utilities}
