% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_functions.R
\name{stats_META}
\alias{stats_META}
\title{Extract the META info from a vcf}
\usage{
stats_META(vcf, vcf_file = "vcf", ...)
}
\arguments{
\item{vcf}{a vcfR object}

\item{vcf_file}{names fro plotting if opt$plot = TRUE}

\item{...}{options from initialise_option()}
}
\value{
list
\item{META }{META data extracted} 
\item{Description_META }{description taken from the vcf header} 
\item{INFO_per_snp }{META data per SNPs} 
\item{CI_Stats }{Statistic for each META data}
}
\description{
Extract the META info from a vcfR object and can report plot.
}
\examples{

opt <- parse_args(OptionParser(option_list=option_list))
opt$plot <- TRUE
stats_META("my.vcf", vcf_file="Plot_metadata", opt)

}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{~utilities}
