% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_VCF_utilities.R
\name{subsample_filterVCF}
\alias{subsample_filterVCF}
\title{Subsample a filterVCF object}
\usage{
subsample_filterVCF(filtervcf, subsample, missing_allowed = NA,
  gt = NULL)
}
\arguments{
\item{filtervcf}{a filterVCF object}

\item{subsample}{regex pattern to grep the VCF column names. Can be a list of individual separated by "|"}

\item{missing_allowed}{Percentage of missing data allowed in the subsample}

\item{gt}{a genotype matrix extracted using extract_gt_ff(), speeds up the function}
}
\value{
a subsampled filterVCF object
}
\description{
Extract a subsample using a list of individuals and recalculate SNPs, invariant and fixed sites
}
\examples{

sub_filterVCF  <- subsample_filterVCF(filterVCF_object, subsample = "ind1|ind3|ind234")

}
\author{
~~Benjamin Laenen~~
}
\keyword{~filterVCF}
