% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_popstat.R
\name{tajimaD_test}
\alias{tajimaD_test}
\title{Tajima's D test}
\usage{
tajimaD_test(d, n, S)
}
\arguments{
\item{d}{numerator of Tajima's D : pi - thetaW}

\item{n}{number of chromosome sampled}

\item{S}{number of segragating sites}
}
\value{
A list with three numeric values:
\item{D }{Tajima's D statistic.} 
\item{Pval.normal	 }{the p-value assuming that D follows a normal distribution with mean zero and variance one.} 
\item{Pval.beta }{the p-value assuming that D follows a beta distribution after rescaling on [0, 1] (Tajima, 1989).}
}
\description{
Function accomodated from pegas. Used internally in get_theta_pi_SFS() and Popstat()
}
\details{
Paradis, E. (2010). pegas: an R package for population genetics with an integrated–modular approach. Bioinformatics, 26(3), 419-420.
}
\note{
~~further notes~~
}
\examples{


}
\references{
Tajima, F. (1989) Statistical method for testing the neutral mutation hypothesis by DNA polymorphism. Genetics, 123, 595–595.
}
\author{
~~Benjamin Laenen~~
}
\keyword{~kwd1}
\keyword{~kwd2}
