% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_convert.R
\name{vcf2Grange}
\alias{vcf2Grange}
\title{Tranform a vcfR object to a GRanges object}
\usage{
vcf2Grange(vcf, metadata = NULL, ...)
}
\arguments{
\item{vcf}{A vcfR object}

\item{metadata}{Should extra information be stored in the GRanges object. Default is NULL, only position are recorded. "present" to get the presence or missing of genotype. "genotype" to get the genotype matrix}

\item{...}{list of options in the form a list object: opt. Only opt$verbose
is used here. Default is true.}
}
\value{
GRanges
}
\description{
Take a vcfR object and transform to genomic ranges GRanges object. ' Some
element can be store in the metadata of the GRanges as S4Vectors. If metadata
= "present" only the the pressence or missing of genotype is recorded. If
metadata = "genotype" is stored in a dataframe in the metadata.
}
\examples{

vcfGrange <- vcf2Grange(vcf = vcfR)
vcfGrange_present <- vcf2Grange(vcf = vcfR, metadata = "present")
vcfGrange_genotype <- vcf2Grange(vcf = vcfR, metadata = "genotype")

elementMetadata(vcfGrange)
elementMetadata(vcfGrange_present)
elementMetadata(vcfGrange_genotype)

}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},

Other utilities VCFutilities: \code{\link{extract_gt_ff}}
}
\author{
~~Benjamin Laenen~~
}
\concept{utilities VCFutilities}
\keyword{~utilities}
