% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_popstat.R
\name{vcf_freq2sweepfinder}
\alias{vcf_freq2sweepfinder}
\title{Create input for Sweepfinder2}
\usage{
vcf_freq2sweepfinder(vcf_freq, output_file = "sweepfinder.input",
  invariant = TRUE)
}
\arguments{
\item{vcf_freq}{a vcf_frequency object obtain using get_frequency()}

\item{output_file}{= "sweepfinder.input" <Name for the output file}

\item{invariant}{= TRUE include the invariant in the sweepfind2 input? The program can be run with or without.}
}
\description{
This function create input for the program sweepfinder2 from a vcf_freq
object obtain using get_frequency()
}
\details{
# The input for sweepfinder looks like this
# position x   n   folded
# 460000   9   100 0
# 460010   100 100 0
# 460210   30  78  1
# 463000   094 0
}
\examples{

freq <- get_frequency(filterVCF_object)
vcf_freq2sweepfinder(freq, output_file = "sweepfinder_all_sites.input") 
vcf_freq2sweepfinder(freq, output_file = "sweepfinder_snps.input",  invariant = FALSE) 

}
\references{

}
\author{
~~Benjamin Laenen~~
}
\keyword{~convert}
