% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\docType{class}
\name{VIF-class}
\alias{VIF-class}
\title{An S4 class to for VariantInvariantFixed (VIF).}
\description{
An S4 class to for VariantInvariantFixed (VIF).
}
\section{Slots}{

\describe{
\item{\code{vcf_filtered}}{The filtered vcf of class "vcfR",}

\item{\code{vcf_inv_filtered}}{The list of invariant filtered sites of class "GRanges",}

\item{\code{fixed_inv_Grange}}{The list of invariant fixed sites of class"GRanges",}

\item{\code{reference}}{A DNAStringSet object with the reference used to create the object,}

\item{\code{removed_sites}}{All sites removed in snp of class "GRanges",}

\item{\code{removed_sites_inv}}{All sites removed in invariant of class "GRanges",}

\item{\code{Divergence}}{from outgroupd of class "GRanges, structuee is chr pos REF out1 out2 ... outn}

\item{\code{Repolarization}}{key of class "GRanges", produce using either likelohood (Keightley) or parsimony.}

\item{\code{bootstrap}}{List to store bootstrap index to reproduce boostrapped dataset quickly,}

\item{\code{opt}}{List of option used by the pipeline for VIF,}
}}

