% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_class_methods.R
\name{VIF}
\alias{VIF}
\title{Create a empty VIF object}
\usage{
VIF(...)
}
\value{
VIF object
}
\description{
Create a empty VIF object
}
\examples{

VIF()

}
\references{
put references to the literature/web site here
}
\author{
~~Benjamin Laenen~~
}
\keyword{~initiator}
