% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_VCF_utilities.R
\name{windows_from_reference}
\alias{windows_from_reference}
\title{Make windows from a reference}
\usage{
windows_from_reference(reference, vcf = NULL, windows_size = 20000,
  steps = NULL)
}
\arguments{
\item{reference}{A DNAstringSet object or the path to a fasta file}

\item{vcf}{A vcfR or GRanges object to keep only the windows with the same chromosome name as the vcf/GRanges}

\item{windows_size}{size of the wondows in base pair. Default is 20000bp}

\item{steps}{number od base pair for slidding windows. NULL by default gives non-overlapping windows}
}
\value{
A GRanges object with windows
}
\description{
This function creates non-overlapping or slidding windows from a reference.
}
\examples{

reference_file <- "my_species_ref.fasta"

#Read the reference
REF <- readDNAStringSet(reference_file)

#Make windows of 20kb non overlapping
windows1 <- windows_from_reference(reference_file)
windows2 <- windows_from_reference(REF)
identical(windows1, windows2)

#Make windows of 50kb overlapping by 10kb
windows <- windows_from_reference(REF, windows_size = 50000, steps = 10000)


}
\references{

}
\author{
~~Benjamin Laenen~~
 objects to See Also as \code{\link{help}},
}
\keyword{~utilities}
