% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_batchCreator.R
\name{write_batch_script_slurm}
\alias{write_batch_script_slurm}
\title{Write a batch script to run filterVCF slurm job}
\usage{
write_batch_script_slurm(vcf_file, outputdir, ...)
}
\arguments{
\item{vcf_file}{path to the vcf}

\item{outputdir}{output directory to save the results}

\item{...}{list of option produced by initialise_option()}
}
\value{
Write batch job files
}
\description{
This function should not be used directly but is part of the filterVCF.R
pipeline
}
\examples{


}
\references{

}
\seealso{
objects to See Also as \code{\link{help}},
}
\author{
~~Benjamin Laenen~~
}
\keyword{~batchCreator}
