% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVCF_vif_repolarization.R
\name{write_input_Keightley}
\alias{write_input_Keightley}
\title{Write input for est-sfs in a file or internally}
\usage{
write_input_Keightley(ACGT, output = NULL, ...)
}
\arguments{
\item{ACGT}{GRanges output from make_GRange_freq_base_Keightley()}

\item{output}{output file name, if NULL the function only return a character vctor with the input to run est-sfs}
}
\value{
a character vector with the input to run est-sfs
}
\description{
This function create a caracter vector using the output from
make_GRange_freq_base_Keightley() to write it to a file if an output file is
specified or return a character with the input.
}
\examples{

input_est-sfs <- make_GRange_freq_base_Keightley(vif_object)
write_input_Keightley(input_est-sfs)


}
\author{
~~Benjamin Laenen~~
}
\keyword{~est-sfs}
